Fixed compile errors
